refactor: move identity to its own module
